[ML] refactor some nlp tokenizer internals
